+++ R/data.R
@@ -41,7 +41,7 @@
 #' 
 #' All variables are averaged over the 10 minute interval unless otherwise noted.
 #' \describe{
-#'   \item{when}{Timestamp for each measurment set in Eastern Standard Time.}
+#'   \item{when}{Timestamp for each measurement set in Eastern Standard Time.}
 #'   \item{temperature}{average temperature, in Celsius}
 #'   \item{wind_speed}{Wind speed, in meters per second}
 #'   \item{wind_dir}{Wind direction, in degrees}
@@ -73,15 +73,15 @@
 #' 
 #' @description Shapefiles from the MacLeish Field Station. The field station
 #' itself is located at \code{lat = 42.449167}, \code{lon = -72.679389}. These
-#' data contains information about various man-made and natural structures
+#' data contain information about various man-made and natural structures
 #' surrounding the field station. 
 #' 
-#' @details Each of the \code{\link[sp]{Spatial}} objects are projected in 
+#' @details Each of the \code{\link[sf]{sf}} objects are projected in 
 #' \code{epsg:4326} for easy integration with \code{\link[ggmap]{ggmap}} or
 #' \code{\link[leaflet]{leaflet}} objects. 
 #' 
 #' @docType data
-#' @format A \code{list} of \code{\link[sp]{Spatial}} objects, each providing a different layer.
+#' @format A \code{list} of \code{\link[sf]{sf}} objects, each providing a different layer.
 #' 
 #' \describe{
 #'  \item{landmarks}{Landmarks}
@@ -97,13 +97,15 @@
 #'  \item{boundary}{the property boundary}
 #'  \item{research}{research plots}
 #'  \item{soil}{soil deposits used by Amy Rhodes}
+#'  \item{contours_30ft}{30 foot contours}
+#'  \item{contours_3m}{3 meter contours}
 #' }
 #' 
 #' @examples 
 #' names(macleish_layers)
-#' summary(macleish_layers[["buildings"]])
+#' macleish_layers[["buildings"]]
 #' 
-#' if (require(sp)) {
+#' if (require(sf)) {
 #'  plot(macleish_layers[["buildings"]])
 #' }
 #'